requirements.txt
babel==2.15.0
crawl4ai==0.4.247
google-generativeai==0.8.4